+++ Ex2/myBank.h
@@ -3,6 +3,6 @@ void checkBalance(int);
 void deposit(int, double);
 void withdraw(int, double);
 void closeAccount(int);
-void addInterest(double);
+void addInterest(int);
 void printAccounts();
 void closeAllAccounts();
